requirements.txt
# Core Framework
python-telegram-bot==20.7
anthropic==0.71.0

# Task Scheduling
APScheduler==3.10.4

# Database & ORM
SQLAlchemy>=2.0.35
alembic>=1.13.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic>=2.10.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
black==23.11.0
mypy==1.7.1
faker==20.1.0
